Directions and Responses
These are the bot responses that will be doable for the bot. They are randomly selected as seen from the 31st line of code.  The function as it stands is completing the first assignment reqirement of creating a get_bot_response() function, the parameter of user_response, at least 2 lists with 3 unique responses, strings, and the use of the choice () command