Conditional Statements
These are the conditional statements and use of choice. The choice () returns a randomly selected element. In our "if/elif/else" statements, if the user_response is one of our desired choices, such as chevy, we are returning a randomly selected string response from our previously defined lists. This will create the body of our coding and what the user understands as the underlying components of a "Chat Bot"